// src/ui/input/Input.d.ts
import { InputHTMLAttributes } from "react";
export declare const Input: import("react").ForwardRefExoticComponent<{
    error?: string;
    inputType?: "default" | "reveal" | "search";
    label?: string;
    onSearchClick?: () => void;
} & InputHTMLAttributes<HTMLInputElement> & import("react").RefAttributes<HTMLInputElement>>;
//# sourceMappingURL=Input.d.ts.map